Extract step logic to concern
