+++ app/views/external_users/claims/supporting_documents/_existing.html.haml
@@ -1,8 +1,6 @@
 - if @claim.documents.count > 0
   %h2.bold-medium
     = t('.previously_uploaded_docs')
-    .form-hint
-    = t('.supporting_evidence_docs_hint')
 
   %table.previously-uploaded
     %thead
